refactor: Extract ESpecialTaskCategory enum for depencency loop breaking.
