+++ ydb/core/config/validation/ya.make
@@ -14,6 +14,7 @@ PEERDIR(
     ydb/core/blobstorage/base
     ydb/core/protos
     ydb/core/formats/arrow/serializer
+    ydb/core/tx/conveyor_composite/common
     library/cpp/protobuf/json
 )
