+++ ydb/core/tx/conveyor_composite/common/ya.make
@@ -1,0 +1,9 @@
+LIBRARY()
+
+SRCS(
+    category.h
+)
+
+GENERATE_ENUM_SERIALIZATION(category.h)
+
+END()
