+++ ydb/core/tx/conveyor_composite/usage/common.h
@@ -1,17 +1,10 @@
 #pragma once
+#include <ydb/core/tx/conveyor_composite/common/category.h>
 #include <ydb/core/tx/conveyor/usage/abstract.h>
 
 namespace NKikimr::NConveyorComposite {
 using ITask = NConveyor::ITask;
 class TCPULimitsConfig;
-
-enum class ESpecialTaskCategory {
-    Insert = 0 /* "insert" */,
-    Compaction = 1 /* "compaction" */,
-    Normalizer = 2 /* "normalizer" */,
-    Scan = 3 /* "scan" */,
-    Deduplication = 4 /* "deduplication" */
-};
 
 class TProcessGuard: TNonCopyable {
 private:
