+++ ydb/core/tx/conveyor_composite/usage/ya.make
@@ -8,12 +8,11 @@ SRCS(
 )
 
 PEERDIR(
+    ydb/core/tx/conveyor_composite/common
     ydb/library/actors/core
     ydb/services/metadata/request
     ydb/core/protos
     ydb/core/tx/conveyor/usage
 )
-
-GENERATE_ENUM_SERIALIZATION(common.h)
 
 END()
